完成消息存储模块,最后就差把历史消息显示在前端.多表连接查询的时候应该注意mapper接口不应该递归调用，这里解释不清，就像GroupInfo对象里会包含list<Belong>对象， 而Belong对象包含了GroupInfo对象，这样我查询一个GroupInfo这样下面会无限嵌套，但是内存却不会出现问题

+++ src/main/java/model/po/GroupInfo.java
@@ -9,7 +9,7 @@ import java.util.List;
  */
 public class GroupInfo {
 
-    private long groupId;
+    private Integer groupId;
     private String groupName;
     private String groupAvatarUrl;
     private List<Belong> members;
@@ -26,11 +26,11 @@ public class GroupInfo {
         this.messageList = messageList;
     }
 
-    public long getGroupId() {
+    public Integer getGroupId() {
         return groupId;
     }
 
-    public void setGroupId(long groupId) {
+    public void setGroupId(Integer groupId) {
         this.groupId = groupId;
     }
